New Function: Returns the true FCS FD for Valve Position FDs by lookup
Intended to help fix/reindex old data sets. Many old data files combined valve state and position. A "state" FD String is parsed for the find number and a pre-built cell array of valid Valve Position FDs is checked for a match.

The correct value is returned. If it is not found, the function attempts to replace "State" with "Position Mon"