Update: partial fix for time axis label spacing.
The issue is with the discrepancy between label size on the screen and label size on the printed PDF. Temporarily setting the spacing calculation to a larger value to force bigger spacing by default.

Users will need to adjust the size of their graph windows and refresh the axes until a permanent solution is found.